Exctracting all reputation logic from auth
