+++ webapp/src/components/Auth/useAuthProvider.ts
@@ -8,84 +8,16 @@ interface AuthState {
   login: (params: LoginParams) => void
   logout: () => void
   profile: Profile | null
-  reputation: Reputation | undefined
 }
-
-const baseRep = {
-  boosters: {
-    hasVotecasterNFTPass: false,
-    hasVotecasterLaunchNFT: false,
-    isVotecasterAlphafrensFollower: false,
-    isVotecasterFarcasterFollower: false,
-    isVocdoniFarcasterFollower: false,
-    votecasterAnnouncementRecasted: false,
-    hasKIWI: false,
-    hasDegenDAONFT: false,
-    hasHaberdasheryNFT: false,
-    has10kDegenAtLeast: false,
-    hasTokyoDAONFT: false,
-    has5ProxyAtLeast: false,
-    hasProxyStudioNFT: false,
-    hasNameDegen: false,
-    hasFarcasterOGNFT: false,
-  },
-  boostersInfo: {
-    degenAtLeast10kPuntuaction: 0,
-    degenDAONFTPuntuaction: 0,
-    farcasterOGNFTPuntuaction: 0,
-    haberdasheryNFTPuntuaction: 0,
-    kiwiPuntuaction: 0,
-    nameDegenPuntuaction: 0,
-    proxyAtLeast5Puntuaction: 0,
-    proxyStudioNFTPuntuaction: 0,
-    tokyoDAONFTPuntuaction: 0,
-    vocdoniFarcasterFollowerPuntuaction: 0,
-    votecasterAlphafrensFollowerPuntuaction: 0,
-    votecasterAnnouncementRecastedPuntuaction: 0,
-    votecasterFarcasterFollowerPuntuaction: 0,
-    votecasterLaunchNFTPuntuaction: 0,
-    votecasterNFTPassPuntuaction: 0,
-  },
-  activityPoints: {
-    followersPoints: 0,
-    createdElectionsPoints: 0,
-    castVotesPoints: 0,
-    participationsPoints: 0,
-    communitiesPoints: 0,
-  },
-  activityCounts: {
-    followersCount: 0,
-    createdElectionsCount: 0,
-    castVotesCount: 0,
-    participationsCount: 0,
-    communitiesCount: 0,
-  },
-  activityInfo: {
-    maxCastedReputation: 0,
-    maxCommunityReputation: 0,
-    maxElectionsReputation: 0,
-    maxFollowersReputation: 0,
-    maxReputation: 0,
-    maxVotesReputation: 0,
-  },
-  totalReputation: 0,
-  totalPoints: 0,
-}
-
-export type Reputation = typeof baseRep
 
 type LoginParams = {
   profile: Profile
   bearer: string
-  reputation: Reputation
 }
 
 export const useAuthProvider = (): AuthState => {
   const [bearer, setBearer] = useState<string | null>(localStorage.getItem('bearer'))
   const [profile, setProfile] = useState<Profile | null>(JSON.parse(localStorage.getItem('profile') || 'null'))
-  const [reputation, setReputation] = useState<Reputation | undefined>(
-    JSON.parse(localStorage.getItem('reputation') || '{}')
-  )
 
   const bearedFetch = useCallback(
     async (input: RequestInfo, init: RequestInit = {}) => {
@@ -107,11 +39,6 @@ export const useAuthProvider = (): AuthState => {
     [bearer]
   )
 
-  const storeReputation = ({ reputation }: { reputation: Reputation }) => {
-    setReputation(reputation)
-    localStorage.setItem('reputation', JSON.stringify(reputation))
-  }
-
   // if no bearer but profile, logout
   useEffect(() => {
     if (!bearer && !!profile) {
@@ -132,36 +59,29 @@ export const useAuthProvider = (): AuthState => {
 
         return response.json()
       })
-      // update reputation
-      .then(storeReputation)
       // network errors or other issues
       .catch(() => {
         logout()
       })
   }, [])
 
-  const login = useCallback(({ profile, bearer, reputation }: LoginParams) => {
+  const login = useCallback(({ profile, bearer }: LoginParams) => {
     localStorage.setItem('bearer', bearer)
     localStorage.setItem('profile', JSON.stringify(profile))
-    localStorage.setItem('reputation', JSON.stringify(reputation))
     setBearer(bearer)
     setProfile(profile)
-    setReputation(reputation)
   }, [])
 
   const logout = useCallback(() => {
     localStorage.removeItem('bearer')
     localStorage.removeItem('profile')
-    localStorage.removeItem('reputation')
     setBearer(null)
     setProfile(null)
-    setReputation(undefined)
   }, [])
 
   return {
-    isAuthenticated: !!bearer && !!profile && !!reputation,
+    isAuthenticated: !!bearer && !!profile,
     profile,
-    reputation,
     login,
     logout,
     bearer,
