+++ webapp/src/components/Layout/Navbar.tsx
@@ -2,6 +2,7 @@ import { Avatar, Box, Flex, Heading, HStack, Icon, IconButton, Stack, useDisclos
 import { GiHamburgerMenu } from 'react-icons/gi'
 import { IoClose } from 'react-icons/io5'
 import { Link } from 'react-router-dom'
+import { useReputation } from '~components/Reputation/useReputation'
 import { ReputationProgress } from '../Auth/Reputation'
 import { SignInButton } from '../Auth/SignInButton'
 import { useAuth } from '../Auth/useAuth'
@@ -44,7 +45,8 @@ const links: NavbarLink[] = [
 
 export const Navbar = () => {
   const { isOpen, onOpen, onClose } = useDisclosure()
-  const { isAuthenticated, profile, reputation } = useAuth()
+  const { isAuthenticated, profile } = useAuth()
+  const { reputation } = useReputation()
 
   return (
     <Box px={{ base: 3 }}>
