+++ webapp/src/pages/points.tsx
@@ -1,8 +1,8 @@
 import { ReputationTable } from '~components/Auth/Reputation'
-import { useAuth } from '~components/Auth/useAuth'
+import { useReputation } from '~components/Reputation/useReputation'
 
 const Points = () => {
-  const { reputation } = useAuth()
+  const { reputation } = useReputation()
 
   // this is a restricted route, reputation should be there
   if (!reputation) return null
